Fix e-matching in basic.ml
This is a fix to the bug I discovered in basic.ml's e-matching
algorithm. The issue is that for a given enode/eclass, there can be
multiple valid substitutions for the same pattern. Thus, enode_matches
and match_in should return a list/iter of substitutions rather than an
option.

Note that generic.ml has the same e-matching algorithm and probably
needs to be fixed as well.